Move getArgsWithUpdatedTransformFile to utils
